[nextest-runner] reorganize reporter code
Move extraction code into its own file, and move some files around.
